requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
discord.py==1.6.0
firebase-admin==5.0.0
google-api-core==1.28.0
google-api-python-client==2.5.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
gTTS==2.2.2
httplib2==0.19.1
idna==2.10
msgpack==1.0.2
multidict==5.1.0
packaging==20.9
proto-plus==1.18.1
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.4
yarl==1.6.3
youtube-dl==2021.4.26
expiringdict==1.2.1
